Create markdown pages manually
