+++ gatsby-browser.js
@@ -1,0 +1,4 @@
+import "./src/styles/variables.css";
+import "./src/styles/basic.css";
+import "./src/styles/layout.css";
+import "./src/styles/special.css";
