+++ gatsby-config.js
@@ -35,17 +35,17 @@ module.exports = {
               maxWidth: 800,
             },
           },
-          {
-            resolve: `@raae/gatsby-remark-oembed`,
-            options: {
-              providers: {
-                // Important to exclude providers
-                // that adds js to the page.
-                // If you do not need them.
-                exclude: ["Reddit", "Instagram", "Twitter", "Flickr"],
-              },
-            },
-          },
+          // {
+          //   resolve: `@raae/gatsby-remark-oembed`,
+          //   options: {
+          //     providers: {
+          //       // Important to exclude providers
+          //       // that adds js to the page.
+          //       // If you do not need them.
+          //       exclude: ["Reddit", "Instagram", "Twitter", "Flickr"],
+          //     },
+          //   },
+          // },
           {
             resolve: `gatsby-remark-responsive-iframe`,
           },
