+++ gatsby-node.js
@@ -1,3 +1,4 @@
+const path = require(`path`);
 const { createFilePath } = require("gatsby-source-filesystem");
 
 exports.onCreateNode = ({ node, actions, getNode }) => {
@@ -14,18 +15,59 @@ exports.onCreateNode = ({ node, actions, getNode }) => {
   }
 };
 
-exports.onCreatePage = ({ page, actions, getNode }) => {
-  const { deletePage } = actions;
-  const node = getNode(page.context?.id);
+exports.createPages = async ({ graphql, actions }) => {
+  const { createPage } = actions;
 
-  // Delete the page if its of type Markdown
-  // and there is "/_" in its slug
-  if (node?.internal.type === "MarkdownRemark") {
-    if (node.fields.slug.includes("/_")) {
-      deletePage(page);
+  const { data } = await graphql(`
+    query loadPagesQuery {
+      allMarkdownRemark {
+        nodes {
+          id
+          fields {
+            slug
+          }
+        }
+      }
     }
-  }
+  `);
+
+  data.allMarkdownRemark.nodes.forEach((node) => {
+    let slug = node.fields.slug;
+    let template = path.resolve(`./src/templates/default.js`);
+
+    if (slug.includes("/_")) {
+      // Skip markdown files containing "/_"
+      return;
+    }
+
+    if (node.fields.slug.includes("queen-emails")) {
+      slug = node.fields.slug.replace("queen-emails", "emails");
+      template = path.resolve(`./src/templates/queen-email.js`);
+    }
+
+    createPage({
+      path: slug,
+      component: template,
+      ownerNodeId: node.id,
+      context: {
+        id: node.id,
+      },
+    });
+  });
 };
+
+// exports.onCreatePage = ({ page, actions, getNode }) => {
+//   const { deletePage } = actions;
+//   const node = getNode(page.context?.id);
+
+//   // Delete the page if its of type Markdown
+//   // and there is "/_" in its slug
+//   if (node?.internal.type === "MarkdownRemark") {
+//     if (node.fields.slug.includes("/_")) {
+//       deletePage(page);
+//     }
+//   }
+// };
 
 // exports.createSchemaCustomization = ({ actions, schema }) => {
 //   const { createTypes } = actions;
