+++ src/pages/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 
-import Layout from "../templates/default";
 import Seo from "../components/seo";
 
 import NewsletterSection from "../content/newsletter-section";
@@ -12,7 +11,7 @@ import MainMenu from "../content/main-menu";
 
 const IndexPage = (props) => {
   return (
-    <Layout>
+    <>
       <Seo {...props} />
       <main>
         <HomeHeader />
@@ -28,7 +27,7 @@ const IndexPage = (props) => {
           <SocialLinks />
         </nav>
       </footer>
-    </Layout>
+    </>
   );
 };
